+++ src/Cache/CacheInterface.php
@@ -17,4 +17,5 @@ interface CacheInterface {
 
     public function delete($keyName);
 
+    public function getPrefixKeyName();
 }